refactor: features/post-management 공통 로직 분리
