+++ src/features/post-management/list/ui/index.ts
@@ -1,10 +1,9 @@
 export * from "./PaginationControls"
-export * from "./UserInfoDialog"
 export * from "./SearchAndFilters"
 export * from "./SearchInput"
 export * from "./FilterSelect"
 export * from "./FiltersContainer"
-export * from "./PostActionPanel"
+export * from "./PostControlPanel"
 export * from "./PostTableRow"
 export * from "./PostTable"
 export * from "./PostManagerList"
